requirements.txt
# Архивы
patool==1.12

# PDF
PyPDF2==3.0.0
pytesseract==0.3.13
Pillow==10.4.0

# DOCX
python-docx==1.1.0

# FB2 и EPUB (XML)
lxml==4.9.3

# DJVU
djvu==0.7.0

# LLM Client
google-generativeai==0.3.0

# Утилиты
regex==2025.9.10  # для регулярных выражений, если используешь
